FAutomataData

Description

Data structures and types.

Documentation

type AState = String

Represents state in Finite Automata FAutomata

type ASymbol = String

Represents symbol of alphabet in both FAutomata and DFAutomata

type DState = Int

Represents state in Deterministic Finite Automata DFAutomata

data FAutomata

Describes Finite Automata with epsilon transitions that is used as input for this program.

Constructors

FA

Fields

states :: [AState]

All states

alphabet :: [ASymbol]

Alphabet

trans :: [Transition]

Transitions

start :: AState

Initial state

end :: [AState]

End states

data Transition

Describes single transition in Finite Automata FAutomata. It is not used for Deterministic Finite Automata DFAutomata.

Constructors

Trans

Fields

fromState :: AState

Actual state

fromSym :: ASymbol

Input symbol, empty string for epsilon

toState :: AState

Next state

printTransition

Arguments

:: [Transition]
Array of Transition for printing

-> String
Output String

Prints array of Transition in desirable format.

data DFAutomata

Describes Deterministic Finite Automata that is used as output for this program.

Constructors

DFA

Fields

dstates :: [EpsClosure]

All states

dalphabet :: [ASymbol]

Alphabet

dtrans :: [DTransition]

Transitions

dstart :: EpsClosure

Initial state

dend :: [EpsClosure]

End states

data EpsClosure

Single state of Deterministic Finite Automata DFAutomata.

Constructors

ECls

Fields

stateName :: DState

Name of the state

origStates :: [AState]

Represents states in original Finite Automata FAutomata

printECls

Arguments

:: [EpsClosure]
Array of EpsClosure for printing

-> String
Output String

Prints array of states of Deterministic Finite Automata DFAutomata.

data DTransition

Describes single transition in Deterministic Finite Automata DFAutomata. It is not used for Finite Automata FAutomata.

Constructors

DTrans

Fields

dfromState :: EpsClosure

Actual state

dfromSym :: ASymbol

Input symbol, cannot be epsilon

dtoState :: EpsClosure

Next state

printDTransition

Arguments

:: [DTransition]
Array of DTransition for printing

-> String
Output String in desirable format

Prints array DTransition.