Main

Description

Program arguments: options [input_file]

options:

-i - prints input Finite Automata FAutomata

-t - transforms Finite Automata FAutomata with epsilon transitions to Deterministic Finite Automata DFAutomata

input_file:

stdin is used if name of the file is not provided

Documentation

main

Arguments

:: IO ()
Program output

Starting point of the program.